wip: Refactor the tests and complete the commit docs.
