--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "extern/csv-parser"]
-	path = extern/csv-parser
-	url = https://github.com/vincentlaucsb/csv-parser.git
